# models/users/repository/models.go
// Code generated by sqlc. DO NOT EDIT.

package repository

import (
	"database/sql"
	"time"
)

type Client struct {
	FanfitUserID int32
	TempField    sql.NullString
}

type User struct {
	ID                int32
	UserTypeID        int32
	FirstName         string
	LastName          string
	Email             string
	CreatedDate       time.Time
	Username          sql.NullString
	PhoneNo           sql.NullString
	Gender            sql.NullString
	ProfilePicture    sql.NullString
	Bio               sql.NullString
	BackgroundPicture sql.NullString
}

type UserType struct {
	ID          int32
	Description string
}